Move Retire from Portfolio pages to Portfolio
